+++ CHANGELOG.md
@@ -7,6 +7,7 @@ HEAD
 * Fixed `array[idx].as<JsonVariant>()` and `object[key].as<JsonVariant>()`
 * Fixed return value of `JsonObject::set()` (issue #350)
 * Fixed undefined behavior in `Prettyfier` and `Print` (issue #354)
+* Fixed parser that incorrectly rejected floats containing a `+` (issue #349)
 
 v5.6.6
 ------